refactor(socketio-core): 重构 SocketIO 编解码器并更新协议模型
- 移除旧的 SocketIODecoder/SocketIOEncoder（已迁移到 codec/impl/）
- 更新 Namespace 和 SocketPacket 适配新编解码架构
- 更新 SocketIODecoderTest 测试用例

+++ socketio-core/src/main/java/com/ccl/socketio/core/namespace/Namespace.java
@@ -49,7 +49,7 @@ public class Namespace {
         ackCallbacks.put(String.valueOf(ackId), new CopyOnWriteArrayList<>());
     }
 
-    public void triggerAck(int ackId, Object data) {
+    public void triggerAck(Long ackId, Object data) {
         String key = String.valueOf(ackId);
         CopyOnWriteArrayList<Consumer<SocketIOClient>> callbacks = ackCallbacks.remove(key);
         if (callbacks != null) {
